build(timefold): migrer vers 2.5.0, verte sans licence Enterprise
Le diagnostic ne passant plus par SolutionManager.analyze(), la 2.x n'a
plus de verrou fonctionnel : 838 tests, 0 échec, sans licence. Seul le
test de contrat se saute, son oracle étant analyze() lui-même.

La recette OpenRewrite officielle ne suffit pas : elle nettoie des imports
hors sujet et réécrit getConstraintName() en getConstraintRef()
.constraintName(), qui n'existe pas en 2.5. Reste à mesurer le réglage du
solveur, dont les optima datent des internals de la 1.x.

+++ src/main/java/dev/sylvain/planning/service/PlanningService.java
@@ -37,7 +37,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import ai.timefold.solver.core.api.domain.solution.ConstraintWeightOverrides;
-import ai.timefold.solver.core.api.score.buildin.hardmediumsoft.HardMediumSoftScore;
+import ai.timefold.solver.core.api.score.HardMediumSoftScore;
 import ai.timefold.solver.core.api.solver.Solver;
 import ai.timefold.solver.core.api.solver.SolutionManager;
 import ai.timefold.solver.core.api.solver.SolverFactory;
@@ -2738,7 +2738,7 @@ public class PlanningService {
         FeasibilityAnalyzer.FeasibilityReport faisabilite = feasibilityAnalyzer.analyze(
                 solved.getAnimateurs(), distinctStands(solved), distinctCreneaux(solved),
                 solved.getContraintesAdHoc());
-        int hardScore = solved.getScore() == null ? 0 : solved.getScore().hardScore();
+        int hardScore = solved.getScore() == null ? 0 : Math.toIntExact(solved.getScore().hardScore());
         List<ContributionAdHoc> contraintesAdHocEnCause = contributionsAdHoc.values().stream()
                 .sorted(Comparator.comparingInt(ContributionAdHoc::violations).reversed()
                         .thenComparing(ContributionAdHoc::contrainteId))
